+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/expression/predicate/logical/Not.java
@@ -8,6 +8,8 @@ package org.elasticsearch.xpack.esql.expression.predicate.logical;
 
 import org.elasticsearch.common.io.stream.NamedWriteableRegistry;
 import org.elasticsearch.common.io.stream.StreamInput;
+import org.elasticsearch.compute.operator.DriverContext;
+import org.elasticsearch.compute.operator.EvalOperator;
 import org.elasticsearch.xpack.esql.capabilities.TranslationAware;
 import org.elasticsearch.xpack.esql.core.QlIllegalArgumentException;
 import org.elasticsearch.xpack.esql.core.expression.Expression;
@@ -18,6 +20,8 @@ import org.elasticsearch.xpack.esql.core.querydsl.query.Query;
 import org.elasticsearch.xpack.esql.core.tree.NodeInfo;
 import org.elasticsearch.xpack.esql.core.tree.Source;
 import org.elasticsearch.xpack.esql.core.type.DataType;
+import org.elasticsearch.xpack.esql.evaluator.mapper.EvaluatorMapper;
+import org.elasticsearch.xpack.esql.evaluator.predicate.operator.logical.NotEvaluator;
 import org.elasticsearch.xpack.esql.optimizer.rules.physical.local.LucenePushdownPredicates;
 import org.elasticsearch.xpack.esql.planner.TranslatorHandler;
 
@@ -26,7 +30,7 @@ import java.io.IOException;
 import static org.elasticsearch.xpack.esql.core.expression.TypeResolutions.ParamOrdinal.DEFAULT;
 import static org.elasticsearch.xpack.esql.core.expression.TypeResolutions.isBoolean;
 
-public class Not extends UnaryScalarFunction implements Negatable<Expression>, TranslationAware {
+public class Not extends UnaryScalarFunction implements EvaluatorMapper, Negatable<Expression>, TranslationAware {
     public static final NamedWriteableRegistry.Entry ENTRY = new NamedWriteableRegistry.Entry(Expression.class, "Not", Not::new);
 
     public Not(Source source, Expression child) {
@@ -78,6 +82,11 @@ public class Not extends UnaryScalarFunction implements Negatable<Expression>, T
     }
 
     @Override
+    public EvalOperator.ExpressionEvaluator.Factory toEvaluator(ToEvaluator toEvaluator) {
+        return new NotEvaluatorFactory(source(), toEvaluator.apply(field()));
+    }
+
+    @Override
     protected Expression canonicalize() {
         if (field() instanceof Negatable) {
             return ((Negatable) field()).negate().canonical();
@@ -107,5 +116,19 @@ public class Not extends UnaryScalarFunction implements Negatable<Expression>, T
     @Override
     public Query asQuery(LucenePushdownPredicates pushdownPredicates, TranslatorHandler handler) {
         return handler.asQuery(pushdownPredicates, field()).negate(source());
+    }
+
+    record NotEvaluatorFactory(Source source, EvalOperator.ExpressionEvaluator.Factory field)
+        implements
+            EvalOperator.ExpressionEvaluator.Factory {
+        @Override
+        public EvalOperator.ExpressionEvaluator get(DriverContext context) {
+            return new NotEvaluator(source, field.get(context), context);
+        }
+
+        @Override
+        public String toString() {
+            return "NotEvaluator[field=" + field + ']';
+        }
     }
 }
